+++ src/Brownie/HttpClient/HeadList.php
@@ -7,6 +7,7 @@
 
 namespace Brownie\HttpClient;
 
+use Brownie\Util\Exception\MethodNotImplementedException;
 use Brownie\Util\StorageList;
 
 /**
@@ -14,6 +15,26 @@ use Brownie\Util\StorageList;
  */
 class HeadList extends StorageList
 {
+
+    /**
+     * Initializing container.
+     */
+    protected function initList()
+    {
+        $list = array();
+        foreach (explode("\r\n", trim($this->getInitData())) as $headerLine) {
+            if (empty($headerLine)) {
+                $list = array();
+                continue;
+            }
+            if ($this->isSkipHeaderLine($headerLine)) {
+                continue;
+            }
+            $item = $this->createCollectionItem($headerLine);
+            $list[strtolower($item->getName())] = $item;
+        }
+        $this->setList($list);
+    }
 
     /**
      * Returns the SetCookie flag of the HTTP header.
@@ -37,5 +58,33 @@ class HeadList extends StorageList
     protected function isHeaderHTTP($headerLine)
     {
         return ('HTTP/' == substr($headerLine, 0, 5));
+    }
+
+    /**
+     * Returns the HTTP header skip flag.
+     *
+     * @param string    $headerLine     HTTP header line.
+     *
+     * @return bool
+     *
+     * @throws MethodNotImplementedException
+     */
+    protected function isSkipHeaderLine($headerLine)
+    {
+        throw new MethodNotImplementedException('Method not implemented');
+    }
+
+    /**
+     * Creates a data model for the collection.
+     *
+     * @param string    $headerLine     HTTP header line.
+     *
+     * @return mixed
+     *
+     * @throws MethodNotImplementedException
+     */
+    public function createCollectionItem($headerLine)
+    {
+        throw new MethodNotImplementedException('Method not implemented');
     }
 }